@@ -3,25 +3,31 @@ package smartcar.com.getting_started_android_sdk;
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.widget.Button;
 import android.util.Log;
+import android.widget.Button;
 
-import com.smartcar.sdk.*;
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.smartcar.sdk.SmartcarAuth;
+import com.smartcar.sdk.SmartcarCallback;
+import com.smartcar.sdk.SmartcarResponse;
 
 import org.json.JSONException;
 import org.json.JSONObject;
 
 import java.io.IOException;
 
-import okhttp3.*;
+import okhttp3.OkHttpClient;
+import okhttp3.Request;
+import okhttp3.Response;
+
 
 public class MainActivity extends AppCompatActivity {
 
+    private Context appContext;
     private static String CLIENT_ID;
     private static String REDIRECT_URI;
     private static String[] SCOPE;
-    private Context appContext;
     private SmartcarAuth smartcarAuth;
 
     @Override
@@ -31,7 +37,7 @@ public class MainActivity extends AppCompatActivity {
 
         appContext = getApplicationContext();
         CLIENT_ID = getString(R.string.client_id);
-        REDIRECT_URI = "sc" + getString(R.string.client_id) + "://exchange";
+        REDIRECT_URI = getString(R.string.smartcar_auth_scheme) + "://" + getString(R.string.smartcar_auth_host);
         SCOPE = new String[]{"required:read_vehicle_info"};
 
         smartcarAuth = new SmartcarAuth(
@@ -40,59 +46,62 @@ public class MainActivity extends AppCompatActivity {
                 SCOPE,
                 true,
                 new SmartcarCallback() {
-            @Override
-            public void handleResponse(final SmartcarResponse smartcarResponse) {
-
-                final OkHttpClient client = new OkHttpClient();
-
-                // Request can not run on the Main Thread
-                // Main Thread is used for UI and therefore can not be blocked
-                new Thread(new Runnable() {
                     @Override
-                    public void run() {
+                    public void handleResponse(final SmartcarResponse smartcarResponse) {
+                        Log.i("MainActivity", smartcarResponse.getCode());
 
-                        // send request to exchange the auth code for the access token
-                        Request exchangeRequest = new Request.Builder()
-                            // Android emulator runs in a VM, therefore localhost will be the
-                            // emulator's own loopback address
-                            .url(getString(R.string.app_server) + "/exchange?code=" + smartcarResponse.getCode())
-                            .build();
+                        final OkHttpClient client = new OkHttpClient();
 
-                        try {
-                            client.newCall(exchangeRequest).execute();
-                        } catch (IOException e) {}
+                        // Request can not run on the Main Thread
+                        // Main Thread is used for UI and therefore can not be blocked
+                        new Thread(new Runnable() {
+                            @Override
+                            public void run() {
 
-                        // send request to retrieve the vehicle info
-                        Request infoRequest = new Request.Builder()
-                            .url(getString(R.string.app_server) + "/vehicle")
-                            .build();
+                                // send request to exchange the auth code for the access token
+                                Request exchangeRequest = new Request.Builder()
+                                    // Android emulator runs in a VM, therefore localhost will be the
+                                    // emulator's own loopback address
+                                    .url(getString(R.string.app_server) + "/exchange?code=" + smartcarResponse.getCode())
+                                    .build();
 
-                        try {
-                            Response response = client.newCall(infoRequest).execute();
+                                try {
+                                    client.newCall(exchangeRequest).execute();
+                                } catch (IOException e) {
+                                    e.printStackTrace();
+                                }
 
-                            String jsonBody = response.body().string();
-                            JSONObject JObject = new JSONObject(jsonBody);
+                                // send request to retrieve the vehicle info
+                                Request infoRequest = new Request.Builder()
+                                    .url(getString(R.string.app_server) + "/vehicle")
+                                    .build();
 
-                            String make = JObject.getString("make");
-                            String model = JObject.getString("model");
-                            String year = JObject.getString("year");
+                                try {
+                                    Response response = client.newCall(infoRequest).execute();
 
-                            Intent intent = new Intent(appContext, DisplayInfoActivity.class);
-                            intent.putExtra("INFO", make + " " + model + " " + year);
-                            startActivity(intent);
-                        } catch (IOException e) {
-                            e.printStackTrace();
-                        } catch (JSONException e) {
-                            e.printStackTrace();
-                        }
-                    }
-                }).start();
+                                    String jsonBody = response.body().string();
+                                    JSONObject JObject = new JSONObject(jsonBody);
+
+                                    String make = JObject.getString("make");
+                                    String model = JObject.getString("model");
+                                    String year = JObject.getString("year");
+
+                                    Intent intent = new Intent(appContext, DisplayInfoActivity.class);
+                                    intent.putExtra("INFO", make + " " + model + " " + year);
+                                    startActivity(intent);
+                                } catch (IOException e) {
+                                    e.printStackTrace();
+                                } catch (JSONException e) {
+                                    e.printStackTrace();
+                                }
+                            }
+                        }).start();
 
 
             }
         });
 
-        Button connectButton = (Button) findViewById(R.id.connect_button);
+        Button connectButton = findViewById(R.id.connect_button);
 
         smartcarAuth.addClickHandler(appContext, connectButton);
     }
